Camera: partially switch to 32-bit usage

+++ services/camera/libcameraservice/device1/CameraHardwareInterface.h
@@ -448,7 +448,7 @@ private:
     int mPreviewWidth;
     int mPreviewHeight;
     int mPreviewFormat;
-    uint64_t mPreviewUsage;
+    int mPreviewUsage;
     int mPreviewSwapInterval;
     android_native_rect_t mPreviewCrop;
